Back End Code/requirements.txt
alembic==1.7.5
astroid==2.9.0
autopep8==1.6.0
bcrypt==3.2.0
cffi==1.15.0
click==8.0.3
colorama==0.4.4
cryptography==36.0.0
Flask==2.0.2
Flask-Bcrypt==0.7.1
Flask-Cors==3.0.10
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
isort==5.10.1
itsdangerous==2.0.1
Jinja2==3.0.3
lazy-object-proxy==1.6.0
Mako==1.1.6
MarkupSafe==2.0.1
mccabe==0.6.1
platformdirs==2.4.0
pycodestyle==2.8.0
pycparser==2.21
PyJWT==2.3.0
pylint==2.12.1
six==1.16.0
SQLAlchemy==1.4.27
toml==0.10.2
typing_extensions==4.0.1
uuid==1.30
Werkzeug==2.0.2
wrapt==1.13.3
